[SYCL] PIMPL refactor
Refactor the command_graph and node classes so that
we interface with the impl types rather than
user exposed types, and just the interface lives in the
public facing headers.

This change also means we can use a `.cpp` file for implementation
code rather than being header only.

The motivation for these changes was trying to get graph submission
through a handler, at which point only the `sycl::detail::queue_impl` class
is available rather than `sycl::queue`
