--- sycl/include/sycl/ext/oneapi/experimental/graph_defines.hpp
@@ -1,28 +1,0 @@
-//==--------- graph_defines.hpp --- SYCL graph extension
-//---------------------------==//
-//
-// Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
-// See https://llvm.org/LICENSE.txt for license information.
-// SPDX-License-Identifier: Apache-2.0 WITH LLVM-exception
-//
-//===----------------------------------------------------------------------===//
-
-#pragma once
-
-namespace sycl {
-__SYCL_INLINE_VER_NAMESPACE(_V1) {
-namespace ext {
-namespace oneapi {
-namespace experimental {
-enum class graph_state {
-  modifiable,
-  executable,
-};
-
-/// Forward declaration for command_graph
-template <graph_state State> class command_graph;
-}
-} // namespace oneapi
-} // namespace ext
-} // __SYCL_INLINE_VER_NAMESPACE(_V1)
-} // namespace sycl
